split user class to other files
